[HAX] Try to deal with broken F-Droid repo
* The directory may exist but not contain the index.
* The index may exist but be corrupt.

This code is horrible.
It's ugly, it's hacky.
And the reason is that it's 3AM, and the rest of the code
to handle app install waiting has never been clean, and this is
showing badly now - it just can't handle failure gracefully.

Let's rewrite that some other day, soon, but till then this does work,
and let's use it till bringup is done