Fix bug where selected apps to install would get unselected
When views are recycled, a listener for checked state changes is already
set. When setting the view to the item's checked state this listener
would get called, messing with the item's checked state.

The fix is to remove the listener before changing the view's state to
match the item state.